opendal_service_s3

Struct S3

pub struct S3 { /* private fields */ }

Aws S3 and compatible services (including minio, digitalocean space, Tencent Cloud Object Storage(COS) and so on) support. For more information about s3-compatible services, refer to Compatible Services.

Capabilities

Depending on its configuration and the backing system, this service can expose:

create_dir

stat

read

write

delete

list

copy

rename

presign

Inspect the effective capability set with [opendal_core::Operator::info] and [opendal_core::OperatorInfo::capability] after building an operator.

Configuration

Use crate::S3Config for serializable configuration and this builder’s methods for direct construction. The field and method documentation defines accepted values, defaults, and environment interaction.

Temporary security credentials

OpenDAL now provides support for S3 temporary security credentials in IAM.

The way to take advantage of this feature is to build your S3 backend with Builder::session_token.

But OpenDAL will not refresh the temporary security credentials, please keep in mind to refresh those credentials in time.

Server Side Encryption

OpenDAL provides full support of S3 Server Side Encryption(SSE) features.

The easiest way to configure them is to use helper functions like

SSE-KMS: server_side_encryption_with_aws_managed_kms_key

SSE-KMS: server_side_encryption_with_customer_managed_kms_key

SSE-S3: server_side_encryption_with_s3_key

SSE-C: server_side_encryption_with_customer_key

If those functions don’t fulfill need, low-level options are also provided:

Use service managed kms key

server_side_encryption="aws:kms"

Use customer provided kms key

server_side_encryption="aws:kms"

server_side_encryption_aws_kms_key_id="your-kms-key"

Use S3 managed key

server_side_encryption="AES256"

Use customer key

server_side_encryption_customer_algorithm="AES256"

server_side_encryption_customer_key="base64-of-your-aes256-key"

server_side_encryption_customer_key_md5="base64-of-your-aes256-key-md5"

After SSE have been configured, all requests send by this backed will attach those headers.

Reference: Protecting data using server-side encryption

Example

Via Builder

Basic Setup

use opendal_core::Operator;
use opendal_core::Result;
use opendal_service_s3::S3;

#[tokio::main]
async fn main() -> Result<()> {
    // Create s3 backend builder.
    let mut builder = S3::default()
      // Set the root for s3, all operations will happen under this root.
      //
      // NOTE: the root must be absolute path.
      .root("/path/to/dir")
      // Set the bucket name. This is required.
      .bucket("test")
      // Set the region. This is required for some services, if you don't care about it, for example Minio service, just set it to "auto", it will be ignored.
      .region("us-east-1")
      // Set the endpoint.
      //
      // For examples:
      // - "https://s3.amazonaws.com"
      // - "http://127.0.0.1:9000"
      // - "https://oss-ap-northeast-1.aliyuncs.com"
      // - "https://cos.ap-seoul.myqcloud.com"
      //
      // Default to "https://s3.amazonaws.com"
      .endpoint("https://s3.amazonaws.com")
      // Set the access_key_id and secret_access_key.
      //
      // OpenDAL will try load credential from the env.
      // If credential not set and no valid credential in env, OpenDAL will
      // send request without signing like anonymous user.
      .access_key_id("access_key_id")
      .secret_access_key("secret_access_key");

    let op: Operator = Operator::new(builder)?;

    Ok(())
}

S3 with SSE-C

use log::info;
use opendal_core::Operator;
use opendal_core::Result;
use opendal_service_s3::S3;

#[tokio::main]
async fn main() -> Result<()> {
    let mut builder = S3::default()
      .root("/path/to/dir")
      .bucket("test")
      .region("us-east-1")
      .endpoint("https://s3.amazonaws.com")
      .access_key_id("access_key_id")
      .secret_access_key("secret_access_key")
      // Enable SSE-C
      .server_side_encryption_with_customer_key("AES256", "customer_key".as_bytes());

    let op = Operator::new(builder)?;
    info!("operator: {:?}", op);

    // Writing your testing code here.

    Ok(())
}

S3 with SSE-KMS and aws managed kms key

use log::info;
use opendal_core::Operator;
use opendal_core::Result;
use opendal_service_s3::S3;

#[tokio::main]
async fn main() -> Result<()> {
    let mut builder = S3::default()
      // Setup builders
      .root("/path/to/dir")
      .bucket("test")
      .region("us-east-1")
      .endpoint("https://s3.amazonaws.com")
      .access_key_id("access_key_id")
      .secret_access_key("secret_access_key")
      // Enable SSE-KMS with aws managed kms key
      .server_side_encryption_with_aws_managed_kms_key();

    let op = Operator::new(builder)?;
    info!("operator: {:?}", op);

    // Writing your testing code here.

    Ok(())
}

S3 with SSE-KMS and customer managed kms key

use log::info;
use opendal_core::Operator;
use opendal_core::Result;
use opendal_service_s3::S3;

#[tokio::main]
async fn main() -> Result<()> {
    let mut builder = S3::default()
      // Setup builders
      .root("/path/to/dir")
      .bucket("test")
      .region("us-east-1")
      .endpoint("https://s3.amazonaws.com")
      .access_key_id("access_key_id")
      .secret_access_key("secret_access_key")
      // Enable SSE-KMS with customer managed kms key
      .server_side_encryption_with_customer_managed_kms_key("aws_kms_key_id");

    let op = Operator::new(builder)?;
    info!("operator: {:?}", op);

    // Writing your testing code here.

    Ok(())
}

S3 with SSE-S3

use log::info;
use opendal_core::Operator;
use opendal_core::Result;
use opendal_service_s3::S3;

#[tokio::main]
async fn main() -> Result<()> {
    let mut builder = S3::default()
      // Setup builders
      .root("/path/to/dir")
      .bucket("test")
      .region("us-east-1")
      .endpoint("https://s3.amazonaws.com")
      .access_key_id("access_key_id")
      .secret_access_key("secret_access_key")
      // Enable SSE-S3
      .server_side_encryption_with_s3_key();

    let op = Operator::new(builder)?;
    info!("operator: {:?}", op);

    // Writing your testing code here.

    Ok(())
}

S3 with default ACL

use log::info;
use opendal_core::Operator;
use opendal_core::Result;
use opendal_service_s3::S3;

#[tokio::main]
async fn main() -> Result<()> {
    let mut builder = S3::default()
      // Setup builders
      .root("/path/to/dir")
      .bucket("test")
      .region("us-east-1")
      .endpoint("https://s3.amazonaws.com")
      .access_key_id("access_key_id")
      .secret_access_key("secret_access_key")
      // Enable public-read ACL
      .default_acl("public-read");

    let op = Operator::new(builder)?;
    info!("operator: {:?}", op);

    // New objects will be created with public-read ACL

    Ok(())
}

Compatible Services

AWS S3

AWS S3 is the default implementations of s3 services. Only bucket is required.

builder.bucket("<bucket_name>");

Alibaba Object Storage Service (OSS)

OSS is a s3 compatible service provided by Alibaba Cloud.

To connect to OSS, we need to set:

endpoint: The endpoint of oss, for example: https://oss-cn-hangzhou.aliyuncs.com

bucket: The bucket name of oss.

OSS provide internal endpoint for used at alibabacloud internally, please visit OSS Regions and endpoints for more details.

OSS only supports the virtual host style, users could meet errors like:

<?xml version="1.0" encoding="UTF-8"?>
<Error>
 <Code>SecondLevelDomainForbidden</Code>
 <Message>The bucket you are attempting to access must be addressed using OSS third level domain.</Message>
 <RequestId>62A1C265292C0632377F021F</RequestId>
 <HostId>oss-cn-hangzhou.aliyuncs.com</HostId>
</Error>

In that case, please enable virtual host style for requesting.

builder.endpoint("https://oss-cn-hangzhou.aliyuncs.com");
builder.region("<region>");
builder.bucket("<bucket_name>");
builder.enable_virtual_host_style();

Minio

minio is an open-source s3 compatible services.

To connect to minio, we need to set:

endpoint: The endpoint of minio, for example: http://127.0.0.1:9000

region: The region of minio. If you don’t care about it, just set it to “auto”, it will be ignored.

bucket: The bucket name of minio.

builder.endpoint("http://127.0.0.1:9000");
builder.region("<region>");
builder.bucket("<bucket_name>");

Oracle Cloud Infrastructure Object Storage

Oracle Cloud Infrastructure (OCI) Object Storage provides an Amazon S3 Compatibility API.

To connect with path-style requests, set:

endpoint: The endpoint containing the Object Storage namespace and OCI region, for example: https://<namespace>.compat.objectstorage.<region>.oci.customer-oci.com.

region: The OCI region identifier, for example: us-ashburn-1.

bucket: The bucket name.

access_key_id and secret_access_key: The access key and secret key from an OCI Customer Secret Key.

// Accepted endpoint includes "oci.customer-oci.com" and "oraclecloud.com", while "oci.customer-oci.com" is preferred.
builder.endpoint(
    "https://<namespace>.compat.objectstorage.<region>.oci.customer-oci.com",
);
builder.region("<region>");
builder.bucket("<bucket_name>");
builder.access_key_id("<access_key>");
builder.secret_access_key("<secret_key>");

OCI also supports virtual-hosted-style requests. Use the endpoint https://vhcompat.objectstorage.<region>.oci.customer-oci.com and call enable_virtual_host_style() when selecting that form.

OCI Object Storage differs from Amazon S3 in the following ways:

OCI uses an Object Storage namespace instead of Amazon S3’s global bucket namespace. Bucket names must be unique within the namespace according to the configured bucket scope.

OCI organizes resources in compartments. Buckets created through the S3 Compatibility API use the tenancy’s designated compartment, which defaults to the root compartment.

OCI encrypts every object at rest and does not allow clients to enable or disable encryption through the API.

OCI does not support object-level ACLs. OCI IAM policies control access instead, so S3 object ACL settings such as default_acl do not apply.

OCI requires AWS Signature Version 4; Signature Version 2 is not supported.

QingStor Object Storage

QingStor Object Storage is a S3-compatible service provided by QingCloud.

To connect to QingStor Object Storage, we need to set:

endpoint: The endpoint of QingStor s3 compatible endpoint, for example: https://s3.pek3b.qingstor.com

bucket: The bucket name.

Scaleway Object Storage

Scaleway Object Storage is a S3-compatible and multi-AZ redundant object storage service.

To connect to Scaleway Object Storage, we need to set:

endpoint: The endpoint of scaleway, for example: https://s3.nl-ams.scw.cloud

region: The region of scaleway.

bucket: The bucket name of scaleway.

Tencent Cloud Object Storage (COS)

COS is a s3 compatible service provided by Tencent Cloud.

To connect to COS, we need to set:

endpoint: The endpoint of cos, for example: https://cos.ap-beijing.myqcloud.com

bucket: The bucket name of cos.

Tigris

Tigris is a globally distributed S3-compatible object storage service.

Tigris serves data from the region nearest to the requester through a single global endpoint, and charges no egress fees.

To connect to Tigris, we need to set:

endpoint: The endpoint of Tigris: https://t3.storage.dev

region: The region of Tigris. Please set it to auto.

bucket: The bucket name of Tigris.

Credentials are standard S3 access keys, provided via access_key_id/secret_access_key or the AWS_ACCESS_KEY_ID/AWS_SECRET_ACCESS_KEY environment variables. See Tigris S3 SDK docs for details.

builder.endpoint("https://t3.storage.dev");
builder.region("auto");
builder.bucket("<bucket_name>");

https://fly.storage.tigris.dev is a legacy endpoint that still works, but https://t3.storage.dev is the current one.

Tigris buckets support zero-copy snapshots and forks. A forked bucket is a regular bucket, so it can be read and written through OpenDAL like any other — useful for testing against production data without copying it.

Wasabi Object Storage

Wasabi is a s3 compatible service.

Cloud storage pricing that is 80% less than Amazon S3.

To connect to wasabi, we need to set:

endpoint: The endpoint of wasabi, for example: https://s3.us-east-2.wasabisys.com

bucket: The bucket name of wasabi.

Refer to What are the service URLs for Wasabi’s different storage regions? for more details.

Cloudflare R2

Cloudflare R2 provides s3 compatible API.

Cloudflare R2 Storage allows developers to store large amounts of unstructured data without the costly egress bandwidth fees associated with typical cloud storage services.

To connect to r2, we need to set:

endpoint: The endpoint of r2, for example: https://<account_id>.r2.cloudflarestorage.com

bucket: The bucket name of r2.

region: When you create a new bucket, the data location is set to Automatic by default. So please use auto for region.

enable_exact_buf_write: R2 requires the non-tailing parts size to be exactly the same. Please enable this option to avoid the error All non-trailing parts must have the same length.

R2 has the following capability differences from S3:

delete_max_size: R2’s delete objects will return Internal Error if the batch is larger than 700. Please override delete_max_size to 700.

stat_with_override_cache_control, stat_with_override_content_disposition, stat_with_override_content_type: R2 doesn’t support stat with response override queries. Please override them to false.

Google Cloud Storage XML API

Google Cloud Storage XML API provides s3 compatible API.

endpoint: The endpoint of Google Cloud Storage XML API, for example: https://storage.googleapis.com

bucket: The bucket name.

To access GCS via S3 API, please enable features = ["native-tls"] in your Cargo.toml to avoid connection being reset when using rustls. Tracking in https://github.com/seanmonstar/reqwest/issues/1809

Ceph Rados Gateway

Ceph supports a RESTful API that is compatible with the basic data access model of the Amazon S3 API.

For more information, refer: https://docs.ceph.com/en/latest/radosgw/s3/

SeaweedFS

SeaweedFS is an open-source distributed storage system that serves an S3 API from its gateway.

To connect to SeaweedFS, we need to set:

endpoint: The endpoint of the SeaweedFS S3 gateway, for example: http://127.0.0.1:8333

region: The region of SeaweedFS. SeaweedFS ignores it, so set it to us-east-1 when a region is required.

bucket: The bucket name.

builder.endpoint("http://127.0.0.1:8333");
builder.region("us-east-1");
builder.bucket("<bucket_name>");

Credentials are standard S3 access keys. A gateway started without any configured identity accepts anonymous requests, so allow_anonymous also works.

SeaweedFS has the following capability differences from S3:

write_can_append: SeaweedFS doesn’t support appending to an existing object through the S3 API. Please override it to false.

Object versioning is per-bucket and off by default. Enable it with PutBucketVersioning before relying on the version capabilities, or override them to false.

Implementations

impl S3Builder

pub fn root(self, root: &str) -> Self

Set root of this backend.

All operations will happen under this root.

pub fn bucket(self, bucket: &str) -> Self

Set bucket name of this backend.

pub fn endpoint(self, endpoint: &str) -> Self

Set endpoint of this backend.

Endpoint must be full uri, e.g.

AWS S3: https://s3.amazonaws.com or https://s3.{region}.amazonaws.com

Cloudflare R2: https://<ACCOUNT_ID>.r2.cloudflarestorage.com

Aliyun OSS: https://{region}.aliyuncs.com

Tencent COS: https://cos.{region}.myqcloud.com

Minio: http://127.0.0.1:9000

If user inputs endpoint without scheme like “s3.amazonaws.com”, we will prepend “https://” before it.

pub fn region(self, region: &str) -> Self

Region represent the signing region of this endpoint. This is required if you are using the default AWS S3 endpoint.

If using a custom endpoint,

If region is set, we will take user’s input first.

If not, we will try to load it from environment.

pub fn profile(self, profile: &str) -> Self

Set the AWS profile used by the default credential provider chain.

The configured profile takes precedence over the AWS_PROFILE environment variable and applies to shared AWS config and credentials files and SSO.

This setting has no effect when Self::disable_config_load is set or when Self::credential_provider_chain replaces the default chain.

pub fn access_key_id(self, v: &str) -> Self

Set access_key_id of this backend.

If access_key_id is set, we will take user’s input first.

If not, we will try to load it from environment.

pub fn secret_access_key(self, v: &str) -> Self

Set secret_access_key of this backend.

If secret_access_key is set, we will take user’s input first.

If not, we will try to load it from environment.

pub fn role_arn(self, v: &str) -> Self

Set role_arn for this backend.

If role_arn is set, we will use already known config as source credential to assume role with role_arn.

pub fn external_id(self, v: &str) -> Self

Set external_id for this backend.

pub fn role_session_name(self, v: &str) -> Self

Set role_session_name for this backend.

pub fn assume_role_duration_seconds(self, v: u32) -> Self

Set assume_role_duration_seconds for this backend.

pub fn assume_role_session_tags(self, tags: HashMap<String, String>) -> Self

Set assume_role_session_tags for this backend.

pub fn default_storage_class(self, v: &str) -> Self

Set default storage_class for this backend.

Available values:

DEEP_ARCHIVE

GLACIER

GLACIER_IR

INTELLIGENT_TIERING

ONEZONE_IA

OUTPOSTS

REDUCED_REDUNDANCY

STANDARD

STANDARD_IA

pub fn server_side_encryption(self, v: &str) -> Self

Set server_side_encryption for this backend.

Available values: AES256, aws:kms.

Note

This function is the low-level setting for SSE related features.

SSE related options should be set carefully to make them works. Please use server_side_encryption_with_* helpers if even possible.

pub fn server_side_encryption_aws_kms_key_id(self, v: &str) -> Self

Set server_side_encryption_aws_kms_key_id for this backend

If server_side_encryption set to aws:kms, and server_side_encryption_aws_kms_key_id is not set, S3 will use aws managed kms key to encrypt data.

If server_side_encryption set to aws:kms, and server_side_encryption_aws_kms_key_id is a valid kms key id, S3 will use the provided kms key to encrypt data.

If the server_side_encryption_aws_kms_key_id is invalid or not found, an error will be returned.

If server_side_encryption is not aws:kms, setting server_side_encryption_aws_kms_key_id is a noop.

Note

This function is the low-level setting for SSE related features.

SSE related options should be set carefully to make them works. Please use server_side_encryption_with_* helpers if even possible.

pub fn server_side_encryption_customer_algorithm(self, v: &str) -> Self

Set server_side_encryption_customer_algorithm for this backend.

Available values: AES256.

Note

This function is the low-level setting for SSE related features.

SSE related options should be set carefully to make them works. Please use server_side_encryption_with_* helpers if even possible.

pub fn server_side_encryption_customer_key(self, v: &str) -> Self

Set server_side_encryption_customer_key for this backend.

Args

v: base64 encoded key that matches algorithm specified in server_side_encryption_customer_algorithm.

Note

This function is the low-level setting for SSE related features.

SSE related options should be set carefully to make them works. Please use server_side_encryption_with_* helpers if even possible.

pub fn server_side_encryption_customer_key_md5(self, v: &str) -> Self

Set server_side_encryption_customer_key_md5 for this backend.

Args

v: MD5 digest of key specified in server_side_encryption_customer_key.

Note

This function is the low-level setting for SSE related features.

SSE related options should be set carefully to make them works. Please use server_side_encryption_with_* helpers if even possible.

pub fn server_side_encryption_with_aws_managed_kms_key(self) -> Self

Enable server side encryption with aws managed kms key

As known as: SSE-KMS

NOTE: This function should not be used along with other server_side_encryption_with_ functions.

pub fn server_side_encryption_with_customer_managed_kms_key( self, aws_kms_key_id: &str, ) -> Self

Enable server side encryption with customer managed kms key

As known as: SSE-KMS

NOTE: This function should not be used along with other server_side_encryption_with_ functions.

pub fn server_side_encryption_with_s3_key(self) -> Self

Enable server side encryption with s3 managed key

As known as: SSE-S3

NOTE: This function should not be used along with other server_side_encryption_with_ functions.

pub fn server_side_encryption_with_customer_key( self, algorithm: &str, key: &[u8], ) -> Self

Enable server side encryption with customer key.

As known as: SSE-C

NOTE: This function should not be used along with other server_side_encryption_with_ functions.

pub fn session_token(self, token: &str) -> Self

Set temporary credential used in AWS S3 connections

Warning

session token’s lifetime is short and requires users to refresh in time.

pub fn disable_config_load(self) -> Self

Disable config load so that opendal will not load config from environment.

For examples:

envs like AWS_ACCESS_KEY_ID

files like ~/.aws/config

pub fn disable_list_objects_v2(self) -> Self

Disable list objects v2 so that opendal will fall back to the older List Objects V1 to list objects.

By default, OpenDAL uses List Objects V2 to list objects. However, some legacy services do not yet support V2.

pub fn enable_request_payer(self) -> Self

Enable request payer so that OpenDAL will send requests with x-amz-request-payer header.

With this option the client accepts to pay for the request and data transfer costs.

pub fn disable_ec2_metadata(self) -> Self

Disable load credential from ec2 metadata.

This option is used to disable the default behavior of opendal to load credential from ec2 metadata, a.k.a, IMDSv2

pub fn skip_signature(self) -> Self

Skip signature will skip loading credentials and signing requests.

pub fn allow_anonymous(self) -> Self

👎Deprecated since 0.57.0:
Please use skip_signature instead of allow_anonymous

Allow anonymous will allow opendal to send request without signing when credential is not loaded.

pub fn enable_virtual_host_style(self) -> Self

Enable virtual host style so that opendal will send API requests in virtual host style instead of path style.

By default, opendal will send API to https://s3.us-east-1.amazonaws.com/bucket_name

Enabled, opendal will send API to https://bucket_name.s3.us-east-1.amazonaws.com

pub fn disable_stat_with_override(self) -> Self

👎Deprecated since 0.57.0:
S3 stat override capabilities are enabled by default and this option is no longer needed.

Deprecated: S3 stat override capabilities are enabled by default.

pub fn enable_versioning(self, _enabled: bool) -> Self

👎Deprecated since 0.57.0:
S3 versioning capability is enabled by default and this option is no longer needed.

Deprecated: S3 versioning capability is enabled by default.

pub fn credential_provider_chain( self, chain: ProvideCredentialChain<Credential>, ) -> Self

Replace the credential providers with a custom chain.

pub fn batch_max_operations(self, _batch_max_operations: usize) -> Self

👎Deprecated since 0.57.0:
S3 delete batch capability is enabled by default and this option is no longer needed.

Deprecated: S3 delete batch capability is enabled by default.

pub fn delete_max_size(self, _delete_max_size: usize) -> Self

👎Deprecated since 0.57.0:
S3 delete batch capability is enabled by default and this option is no longer needed.

Deprecated: S3 delete batch capability is enabled by default.

pub fn checksum_algorithm(self, checksum_algorithm: &str) -> Self

Set checksum algorithm of this backend. This is necessary when writing to AWS S3 Buckets with Object Lock enabled for example.

Available options:

“crc32c”

“md5”

pub fn disable_write_with_if_match(self) -> Self

👎Deprecated since 0.57.0:
S3 write with If-Match capability is enabled by default and this option is no longer needed.

Deprecated: S3 write with If-Match capability is enabled by default.

pub fn enable_write_with_append(self) -> Self

👎Deprecated since 0.57.0:
S3 append capability is enabled by default and this option is no longer needed.

Deprecated: S3 append capability is enabled by default.

pub async fn detect_region(endpoint: &str, bucket: &str) -> Option<String>

Detect region of S3 bucket.

Args

endpoint: the endpoint of S3 service

bucket: the bucket of S3 service

Return

Some(region) means we detect the region successfully

None means we can’t detect the region or meeting errors.

Notes

We will try to detect region by the following methods.

Match endpoint with given rules to get region

Cloudflare R2

AWS S3

Aliyun OSS

Send a HEAD request to endpoint with bucket name to get x-amz-bucket-region.

Examples

use opendal_service_s3::S3;

let region: Option<String> = S3::detect_region("https://s3.amazonaws.com", "example").await;

Reference

Amazon S3 HeadBucket API

pub fn default_acl(self, acl: &str) -> Self

Set default ACL for new objects.